Change indexed pixel storage and validation
The goal was to make pixel data self-contained. To make that work, we
now split pixel data into `RawPixels` which are extracted from the
Aseprite file and `Pixels` which have been enriched with some non-local
data. Specifically, for indexed data this means that the data includes a
reference to the palette and related data.

The validation calls now mostly change the type of the data making it
clear in the type whether data has been validated or not.