Relax pixel width and height check to determine the ratio
The Aseprite File Format Specification says the following about the pixel
width and pixel height fields:

  BYTE        Pixel width (pixel ratio is "pixel width/pixel height").
              If this or pixel height field is zero, pixel ratio is 1:1
  BYTE        Pixel height

But the read_aseprite() function only considers a 1:1 pixel ratio if both
width and height values are 1, and not when one of these has zero value.

There are some Aseprite editors (such as LibreSprite) that don't set these
fields, even when only supports a 1:1 pixel ratio. But due read_aseprite()
being that strict in checking the width and height, these Aseprite files
can't be read and attempting to do so will fail with the following error:

  UnsupportedFeature("Only pixel width:height ratio of 1:1 supported")